requirements.txt
discord.py
simple_term_menu
